requirements.txt
Flask==2.3.2
textblob==0.17.1
pyspellchecker==0.7.1
rapidfuzz==2.15.1
wordfreq==3.1.1
gunicorn==21.2.0
tensorflow==2.15.0
numpy==1.26.4
opencv-python-headless==4.8.1.78
Pillow==10.1.0
werkzeug==3.0.1
protobuf==4.24.4
itsdangerous==2.1.2
jinja2==3.1.2
